requirements.txt
Flask>=2.3,<4.0
Pillow>=10.0,<12.0

